requirements.txt
streamlit
nltk
scikit-learn
rapidfuzz
sentence-transformers
flashtext
pdfplumber
reportlab
google-generativeai